pyproject.toml
[project]
name = "mailbox-reader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "emlx",
    "pandas",
    "jupyter",
]
